Removed dependency on computed property names The output of the loader was using computed property names for media queries, e.g.:
```
return {
  [theme.breakpoints.down('xs')]: {
    ...
  }
}
```

This required the css-to-mui-loader to be chained with a transpiling JavaScript loader, such as Babel. The loader now outputs the following:

```
var output = {};

output[theme.breakpoints.down('xs')] = {
  ...
};

return output;
```

Now the output does not use computed property names (or any ES2015 JS feature), so it should no longer need to be chained with Babel.